Isolate compile-time code to avoid transitive dependencies
